+++ libs/framework-core/source/ftrack_framework_core/host/__init__.py
@@ -202,6 +202,11 @@ class Host(object):
             self.id, self._verify_plugins_callback
         )
 
+        # Subscribe to sync tool config from the client
+        self.event_manager.subscribe.host_sync_tool_config(
+            self.id, self._sync_tool_config_callback
+        )
+
     def _client_context_change_callback(self, event):
         '''Callback when the client has changed context'''
         context_id = event['data']['context_id']
@@ -291,6 +296,7 @@ class Host(object):
         client_options = event['data']['client_options']
         payload = event['data']['payload']
 
+        # TODO: we should be able to replace this to: tool_config = self.registry.get_one(extension_type='tool_config', reference=tool_config_reference)
         for typed_configs in self.tool_configs.values():
             tool_config = None
             for _tool_config in typed_configs:
@@ -378,3 +384,77 @@ class Host(object):
                 f'correct extensions path: {unregistered_plugins}'
             )
         return unregistered_plugins
+
+    # def augment_tool_config_callback(self, event):
+    #     '''
+    #     Runs the data with the defined engine type of the given *event*
+    #
+    #     Returns result of the engine run.
+    #
+    #     *event* : Published from the client host connection at
+    #     :meth:`~ftrack_framework_core.client.HostConnection.run`
+    #     '''
+    #
+    #     tool_config_reference = event['data']['tool_config_reference']
+    #     section = event['data']['section']
+    #     new_item = event['data']['new_item']
+    #
+    #     if not new_item:
+    #         return
+    #
+    #     tool_config = self.registry.get_one(extension_type='tool_config', reference=tool_config_reference)
+    #
+    #     self.registry.augment_tool_config(
+    #         tool_config=tool_config,
+    #         section=section,
+    #         new_item=new_item
+    #     )
+    #     # TODO: we need now to sync the tool config to the host connection.
+    #
+    #     # Need to unsubscribe to make sure we subscribe again with the new
+    #     # context
+    #     self.event_manager.unsubscribe(self._discover_host_subscribe_id)
+    #     # Reply to discover_host_callback to clients to pass the host information, so we make sure that if a new connection is made, it receives the correct tool_config
+    #     discover_host_callback_reply = partial(
+    #         provide_host_information,
+    #         self.id,
+    #         self.context_id,
+    #         self.tool_configs,
+    #     )
+    #     self._discover_host_subscribe_id = (
+    #         self.event_manager.subscribe.discover_host(
+    #             callback=discover_host_callback_reply
+    #         )
+    #     )
+    #     self.event_manager.publish.host_context_changed(
+    #         self.id, self.context_id
+    #     )
+
+    def _sync_tool_config_callback(self, event):
+        '''
+        Runs the data with the defined engine type of the given *event*
+
+        Returns result of the engine run.
+
+        *event* : Published from the client host connection at
+        :meth:`~ftrack_framework_core.client.HostConnection.run`
+        '''
+
+        tool_config = event['data']['tool_config']
+
+        registered_tool_config = self.registry.get_one(
+            extension_type='tool_config', reference=tool_config['reference']
+        )
+
+        if not registered_tool_config:
+            self.registry.add(
+                tool_config['type'],
+                name=tool_config['name'],
+                extension=tool_config,
+                path="Memory",
+                create_reference=False,
+            )
+        else:
+            registered_tool_config = tool_config
+
+        return registered_tool_config